Refactor routes and controllers: consolidate route definitions and remove unused controller files
